uridiag: Prevent overwrite of manufacturer memory / add options
This update moves the EEPROM user memory from position 6 to position 51.  We cannot overwrite the manufacturer data programmed in memory positions 0 to 50 (CM119B).

Added a new option to dump the EEPROM memory.
Added a new option to print the manufacturer data.
Added a new option to program the manufacturer data for the CM119B.

Documentation added for functions.  Code was reviewed and updated to standards.